Add example msw for a test and rename handlerfns -> handlers
