+++ packages/invoices/src/OutboundShipment/ListView/ListView.stories.tsx
@@ -1,12 +1,26 @@
 import React from 'react';
 import { Route } from 'react-router';
 import { ComponentStory, ComponentMeta } from '@storybook/react';
-// import { handlers } from '@openmsupply-client/mock-server/src/worker/handlers';
-import { StoryProvider, TestingRouter } from '@openmsupply-client/common';
+
+import {
+  StoryProvider,
+  TestingRouter,
+  mockInvoicesQuery,
+} from '@openmsupply-client/common';
 
 import { OutboundShipmentListView } from './ListView';
 
-const handlers: any[] = [];
+const invoicesQuery = mockInvoicesQuery((_, res, ctx) => {
+  return res(
+    ctx.data({
+      invoices: {
+        __typename: 'InvoiceConnector',
+        totalCount: 0,
+        nodes: [],
+      },
+    })
+  );
+});
 
 export default {
   title: 'Page/OutboundShipmentListView',
@@ -26,5 +40,5 @@ const Template: ComponentStory<typeof OutboundShipmentListView> = args => (
 
 export const Primary = Template.bind({});
 Primary.parameters = {
-  msw: handlers,
+  msw: [invoicesQuery],
 };
