+++ packages/mock-server/src/worker/client.ts
@@ -1,4 +1,4 @@
 import { setupWorker, SetupWorkerApi } from 'msw';
-import { Handlers } from './handlerfns';
+import { Handlers } from './handlers';
 
 export const setupMockWorker = (): SetupWorkerApi => setupWorker(...Handlers);
