+++ packages/mock-server/src/worker/server.ts
@@ -1,4 +1,4 @@
 import { setupServer, SetupServerApi } from 'msw/node';
-import { Handlers } from './handlerfns';
+import { Handlers } from './handlers';
 
 export const setupMockServer = (): SetupServerApi => setupServer(...Handlers);
